size_t

+++ tfds/tf_fifo_queue.hpp
@@ -35,7 +35,7 @@ public:
     }
 
     // O(1)
-    int length() const {
+    size_t length() const {
         return list.length();
     }
